commit on master

+++ modules/thirdTransaction.js
@@ -68,18 +68,18 @@ const findByIDinc = await previewStoreSchema.findOne({store:savesecondmodel.to,
         console.log(findByID.type   !== findByIDinc.type !==saver.unit)
 switch (saver.transaction) {
     case "تحويل":
-      if (findByID.type   !== findByIDinc.type !==saver.unit ) 
+      if (findByID.type   && findByIDinc.type ==saver.unit ) 
       {await thirdModel.findByIdAndDelete(saver._id) 
      
-         return  console.log("first error")}
+         return  console.log("error")}
 
-      if (!(findByID  || findByIDinc || (findByID.quantity - saver.quantity) < 0 )) 
+      if (findByID  || findByIDinc || (findByID.quantity - saver.quantity) < 0 ) 
        {await thirdModel.findByIdAndDelete(saver._id) 
         
-            return  console.log("second error")}
+            return  console.log("error")}
         if(findByID.quantity < 0)    {await thirdModel.findByIdAndDelete(saver._id) 
         
-        return  console.log("third error")
+        return  console.log("error")
         }
         const updatedDec = await previewStoreSchema.findByIdAndUpdate(findByID._id,{"$inc":{quantity:- saver.quantity}})
         const updatedInc = await previewStoreSchema.findByIdAndUpdate(findByIDinc._id,{"$inc":{quantity:+ saver.quantity}})
